requirements.txt
click
PyYAML
pytest
pytest-mock